refactor(host): split composeHost into composers
`composeHost` was one 2,600-line function holding every concern of the
runtime in a single scope: 24 mutable `let`s nothing could keep private, a
567-line method table covering seven unrelated domains, and a dependency
graph that existed only as closure capture.

It is now seven composers — settings, account, issues, observation,
calendars, speech, brain — each owning its own state, its own timers, and
the Gateway methods of its domain, each answering `start()` and `stop()`
for exactly what it began. `compose-host.ts` constructs them, links the
five back-edges the concerns genuinely have, merges their method tables,
and keeps the drain. A method two composers claim is a construction
failure rather than a last-writer-wins, and a late reference read before
`link()` has run throws by name.

`RECEIVER_REPORT_KIND` is the vocabulary of `receiver.report`, so it moves
to the protocol beside the method it names. `composeNotebookMemory` and
`transitionVoiceCredential` move to the files their tests were already
named for. No method's semantics, no event's payload, and no ordering the
launch or the quit depends on has changed.
