+++ packages/host/AGENTS.md
@@ -20,6 +20,25 @@ that line: the ipcMain registrations stayed in `apps/desktop/src/main/ipc/`,
 and resolving this Mac's EventKit helper bundle stayed in
 `apps/desktop/src/main/native/`.
 
+## One composition, seven concerns
+
+`composeHost` constructs, links, merges, and starts; it holds no state of a
+concern's own. Each concern is a composer — settings, account, issues,
+observation, calendars, speech, brain — that owns its own mutable state, its
+own timers, and the Gateway methods of its domain, and answers `start()` and
+`stop()` for exactly what it began. The merge folds their method tables into
+one and refuses a method two of them claim, so which concern answers a method
+is checked at construction rather than left to the fold's order.
+`client.bootstrap` is the one method no composer owns: it reads six of them,
+and giving it to any would hand that composer references to the other five.
+
+The concerns depend on each other in both directions in five places — the
+account's capability gate starts the loops whose owners read that gate, the
+calendars hold the speech that reconciles against them — so those edges are
+`link()`'s, listed once in the merge and held in a `LateRef` that throws by
+name when read before `link()` has run. Everything else is a constructor
+argument, in the order the composers are built.
+
 ## One drain, in one place
 
 `Host.stop()` is the whole quit, in the coordinator's fixed order:
