+++ packages/host/src/index.ts
@@ -14,8 +14,9 @@
  * client does not reach yet is added when one arrives.
  */
 export { REJECTED_SUBMISSION } from "./brain/publication.js";
-export { composeHost, type Host, type HostSeams, RECEIVER_REPORT_KIND } from "./compose-host.js";
+export { composeHost, type Host } from "./compose-host.js";
 export type { ConversationOperations } from "./conversation-operations.js";
+export type { HostSeams } from "./host-kernel.js";
 export {
   INTRODUCTION_FADE_MS,
   INTRODUCTION_HANDOFF_READY_MS,
